% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{dof.occuR}
\alias{dof.occuR}
\title{Compute (estimated) degrees of freedom for fitted occupancy model}
\usage{
dof.occuR(obj, each = FALSE)
}
\arguments{
\item{obj}{fitted model object}

\item{each}{default is FALSE, if TRUE then list of degrees of freedom is returned, one fixed parameters, smooth
for psi, and smooth for p. When FALSE, these are summed and total estimated degrees of freedom returned.}
}
\value{
degrees of freedom (see "each" above)
}
\description{
Compute (estimated) degrees of freedom for fitted occupancy model
}
