% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{edf.occuR}
\alias{edf.occuR}
\title{Compute estimated degrees of freedom for a spline}
\usage{
edf.occuR(X, S, lambda, Sn)
}
\arguments{
\item{X}{design matrix for spline}

\item{S}{smoothing matrix for spline}

\item{lambda}{smoothiing parameters}

\item{Sn}{number of parameters per spline}
}
\value{
estimated degrees of freedom
}
\description{
This code was written by David L Miller as part of the CTMCdive R package
and is based on Wood et al. (2017) p 212.
}
