% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{fit_occu}
\alias{fit_occu}
\title{Fit occupancy model}
\usage{
fit_occu(forms, visit_data, site_data, start = NULL, print = TRUE)
}
\arguments{
\item{forms}{list of formulae for psi (occupancy) and p (detection)}

\item{visit_data}{data.table with row for each visit with a "obs" column for detection record}

\item{site_data}{data.table with row for each site x occasion}

\item{start}{named list of starting values for beta_psi and beta_p - fixed effects parameters}

\item{print}{if TRUE then possibly useful info is printed out}
}
\value{
fitted occupied model object
}
\description{
Fit occupancy model
}
