% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{logLik.occuR}
\alias{logLik.occuR}
\title{Log-likelihood for fitted occupancy model}
\usage{
\method{logLik}{occuR}(object, ...)
}
\arguments{
\item{object}{fitted occupancy model from fit_occu}

\item{...}{}
}
\value{
log-likelihood with estimated degrees of freedom as attribute "df"
}
\description{
Log-likelihood for fitted occupancy model
}
