% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_matrices.R
\name{make_matrices}
\alias{make_matrices}
\title{Compute design and smoothing matrices from formulae}
\usage{
make_matrices(forms, visit_data, site_data)
}
\arguments{
\item{forms}{list of formulae for psi and p}

\item{visit_data}{see fit_occu function}

\item{site_data}{see fit_occu function}
}
\value{
list of X = fixed effects design matrix, Z = random effects design matrix,
  and S = smoothing matrix for both psi and p (e.g. X_psi).
  make_smoothing_matrix() returns a list of S (block diagonal smoothing matrix),
  Scols (number of columns for each smoothing matrix),
  bdim (k argument used for each smooth), Snames (names of smoothing parameters).
  Returns NULL if no smooths in model.
  Based on code written by David L Miller in CTMCdive R Package (https://github.com/r-glennie/CTMCdive).
}
\description{
Compute design and smoothing matrices from formulae
}
