% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{make_smoothing_matrix}
\alias{make_smoothing_matrix}
\title{Compute block-diagonal smoothing matrix}
\usage{
make_smoothing_matrix(gm)
}
\arguments{
\item{gm}{value returned by mgcv::gam with fit = FALSE}
}
\value{
list of S (block diagonal smoothing matrix), Scols (number of columns for each smoothing matrix),
bdim (k argument used for each smooth), Snames (names of smoothing parameters). Returns NULL if no smooths in
model.
}
\description{
Based on code written by David L Miller in CTMCdive R Package (https://github.com/r-glennie/CTMCdive).
}
