% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{print.occuR}
\alias{print.occuR}
\title{Print of occuR object}
\usage{
\method{print}{occuR}(obj)
}
\arguments{
\item{obj}{output from fit_occu}
}
\value{
prints a summary
}
\description{
Print of occuR object
}
