% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{summary.occuR}
\alias{summary.occuR}
\title{Summary of occuR object}
\usage{
\method{summary}{occuR}(obj, conf.level = 0.95)
}
\arguments{
\item{obj}{output from fit_occu}
}
\value{
prints a summary
}
\description{
Summary of occuR object
}
