% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{make_matrices}
\alias{make_matrices}
\title{Compute design and smoothing matrices from formulae}
\usage{
make_matrices(forms, visit_data, site_data)
}
\arguments{
\item{forms}{list of formulae for psi and p}

\item{visit_data}{see fit_occu function}

\item{site_data}{see fit_occu function}
}
\value{
list of X = fixed effects design matrix, Z = random effects design matrix,
  and S = smoothing matrix for both psi and p (e.g. X_psi)
}
\description{
Compute design and smoothing matrices from formulae
}
