% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occu_tmb.R
\name{predict.occuR}
\alias{predict.occuR}
\title{Predict from fitted occupancy model}
\usage{
\method{predict}{occuR}(obj, visit_data, site_data, nboot = 0)
}
\arguments{
\item{obj}{fitted model object from fit_occu}

\item{visit_data}{see fit_occu}

\item{site_data}{see fit_occu}

\item{nboot}{number of parametric bootstrap resamples to produce from fitted model}
}
\value{
if nboot = 0 (default), the list of fitted psi and p values; otherwise, list also 
contains matrix for psi and p where each row is a bootstrap resample
}
\description{
Predict from fitted occupancy model
}
